Remove the change of type of subscription id-s

+++ src/app/reporting/Engine.cpp
@@ -446,7 +446,7 @@ CHIP_ERROR Engine::BuildAndSendSingleReportData(ReadHandler * apReadHandler)
 
     if (apReadHandler->IsType(ReadHandler::InteractionType::Subscribe))
     {
-        uint32_t subscriptionId = 0;
+        uint64_t subscriptionId = 0;
         apReadHandler->GetSubscriptionId(subscriptionId);
         reportDataBuilder.SubscriptionId(subscriptionId);
     }
